Moved projector backends to dedicated module
